% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcca_mcmc.R
\name{cmcca_mcmc}
\alias{cmcca_mcmc}
\title{Function for semiparametric CCA model}
\usage{
cmcca_mcmc(Y1, Y2, iter = 1000, burn_in = 0, thin = 1, plot = F, d = NULL)
}
\description{
Function for semiparametric CCA model
}
